requirements.txt
fastapi
aio_pika
uvicorn
python-dotenv
